# frontend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
...
    "/api/v1/admin/scm-connections/{id}" : {
      "get" : {
        "tags" : [ "scm-connection-admin-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ScmConnectionView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "scm-connection-admin-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ScmConnectionUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ScmConnectionView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "scm-connection-admin-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/admin/scm-connections" : {
      "get" : {
        "tags" : [ "scm-connection-admin-controller" ],
        "operationId" : "list_2",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ScmConnectionView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "scm-connection-admin-controller" ],
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ScmConnectionUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ScmConnectionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/scm-connections/{id}/test" : {
      "post" : {
        "tags" : [ "scm-connection-admin-controller" ],
        "operationId" : "test",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repositories" : {
      "get" : {
        "tags" : [ "repository-controller" ],
        "operationId" : "list_4",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageRepositorySummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}" : {
      "get" : {
        "tags" : [ "repository-controller" ],
        "operationId" : "get_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RepositoryDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repositories/{id}/graph" : {
      "get" : {
        "tags" : [ "repo-graph-controller" ],
        "operationId" : "graph",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "level",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "focus",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "includeExternal",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RepoGraph"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/graph/report" : {
      "get" : {
        "tags" : [ "repo-graph-controller" ],
        "operationId" : "report",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RepoGraphReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/graph/export" : {
      "get" : {
        "tags" : [ "repo-graph-controller" ],
        "operationId" : "export_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "format",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "level",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "focus",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "includeExternal",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ScmConnectionUpdate" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "BITBUCKET_DC", "GITHUB", "GIT" ]
          },
          "baseUrl" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "secret" : {
            "type" : "string"
          },
          "includeProjects" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "excludeRepos" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "repositoryUrls" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "includeOwnRepositories" : {
            "type" : "boolean"
          },
          "enabled" : {
            "type" : "boolean"
          }
        }
      },
      "ScmConnectionView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "BITBUCKET_DC", "GITHUB", "GIT" ]
          },
          "baseUrl" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "secretSet" : {
            "type" : "boolean"
          },
          "includeProjects" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "excludeRepos" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "repositoryUrls" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "includeOwnRepositories" : {
            "type" : "boolean"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "lastTestStatus" : {
            "type" : "string"
          },
          "lastTestAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastSyncStatus" : {
            "type" : "string"
          },
          "lastSyncAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastSyncError" : {
            "type" : "string"
          },
          "repositoryCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "PageRepositorySummary" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RepositorySummary"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "RepositorySummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "projectKey" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "defaultBranch" : {
            "type" : "string"
          },
          "lastIndexedCommit" : {
            "type" : "string"
          },
          "lastIndexedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastStatus" : {
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "moduleCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ModuleView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "path" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "string"
          },
          "artifactId" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          },
          "classpathMode" : {
            "type" : "string"
          }
        }
      },
      "RepositoryDetail" : {
        "type" : "object",
        "properties" : {
          "repository" : {
            "$ref" : "#/components/schemas/RepositorySummary"
          },
          "modules" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ModuleView"
            }
          }
        }
      },
...
      "ClassMetrics" : {
        "type" : "object",
        "properties" : {
          "inDegree" : {
            "type" : "integer",
            "format" : "int32"
          },
          "outDegree" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dependents" : {
            "type" : "integer",
            "format" : "int32"
          },
          "entryPoint" : {
            "type" : "boolean"
          },
          "communityId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "communityLabel" : {
            "type" : "string"
          }
        }
      },
      "GraphEdge" : {
        "type" : "object",
        "properties" : {
          "from" : {
            "type" : "string"
          },
          "to" : {
            "type" : "string"
          },
          "weight" : {
            "type" : "integer",
            "format" : "int64"
          },
          "kinds" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
      "GraphNode" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "MODULE", "PACKAGE", "CLASS", "METHOD", "CONSTRUCTOR", "FIELD", "EXTERNAL_REPOSITORY", "EXTERNAL_LIBRARY" ]
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "metrics" : {
            "$ref" : "#/components/schemas/ClassMetrics"
          }
        }
      },
      "RepoGraph" : {
        "type" : "object",
        "properties" : {
          "repositoryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "requestedLevel" : {
            "type" : "string",
            "enum" : [ "MODULE", "PACKAGE", "CLASS", "METHOD" ]
          },
          "level" : {
            "type" : "string",
            "enum" : [ "MODULE", "PACKAGE", "CLASS", "METHOD" ]
          },
          "focus" : {
            "type" : "string"
          },
          "truncated" : {
            "type" : "boolean"
          },
          "suggestion" : {
            "type" : "string"
          },
          "nodes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GraphNode"
            }
          },
          "edges" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GraphEdge"
            }
          }
        }
      },
      "CommunitySummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "label" : {
            "type" : "string"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "CriticalClass" : {
        "type" : "object",
        "properties" : {
          "symbolId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "fqn" : {
            "type" : "string"
          },
          "inDegree" : {
            "type" : "integer",
            "format" : "int32"
          },
          "outDegree" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dependents" : {
            "type" : "integer",
            "format" : "int32"
          },
          "entryPoint" : {
            "type" : "boolean"
          }
        }
      },
      "PackageCycle" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "packages" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "RepoGraphReport" : {
        "type" : "object",
        "properties" : {
          "repositoryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "indexedCommit" : {
            "type" : "string"
          },
          "analyzedCommit" : {
            "type" : "string"
          },
          "analyzedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "stale" : {
            "type" : "boolean"
          },
          "moduleCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "packageCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "classCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dependencyCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "communityCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "criticalClasses" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CriticalClass"
            }
          },
          "communities" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommunitySummary"
            }
          },
          "cycles" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PackageCycle"
            }
          },
          "entryPointCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "entryPointClasses" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
    }
  }
}